+++ rules/aayaan/rule_con40.c
@@ -1,5 +1,5 @@
 /**
- * @file CON40-C.c
+ * @file rule_con40.c
  * @author Aayaan Shaikh
  * @brief demonstrates how to preserve atomicity of atomic value
  * @version 0.1
@@ -17,17 +17,17 @@
  */
 int getSquare(int n)
 {
-  // CON40-C: Do not refer to an atomic variable twice in an expression
-  // avoided refering to atomic var twice in expression
-  // pass atomic var as parameter instead to retain atomic nature
-  return n * n;
+    // CON40-C: Do not refer to an atomic variable twice in an expression
+    // avoided refering to atomic var twice in expression
+    // pass atomic var as parameter instead to retain atomic nature
+    return n * n;
 }
 
 int main(void)
 {
-  atomic_int n = ATOMIC_VAR_INIT(5);
+    atomic_int n = ATOMIC_VAR_INIT(5);
 
-  printf("n^2 is: %d\n", getSquare(n));
+    printf("n^2 is: %d\n", getSquare(n));
 
-  return 0;
+    return 0;
 }
